\name{WilcoxCV-internal}
\alias{wilcox.split.internal}
\alias{wilcox.test2}


\title{Internal WilcoxCV Functions}
\description{
  Internal WilcoxCV functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
